# api-docs.json
{
  "openapi":"3.1.0",
  "info":{
    "title":"Election Handler API",
    "description":"Only for testing purposes",
    "version":"1.0"
  },
  "servers":[
    {
      "url":"http://localhost:8080",
      "description":"Generated server url"
    }
  ],
  "paths":{
    "/api/v1/voting/add-vote":{
      "post":{
        "tags":[
          "Voting"
        ],
        "summary":"Add user voting",
        "operationId":"addVoting",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/VotingRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK"
          }
        }
      }
    },
...
    "/api/v1/voting/data/{voteIdentifier}":{
      "get":{
        "tags":[
          "Voting"
        ],
        "summary":"Get voting data",
        "operationId":"getActiveVotingData",
        "parameters":[
          {
            "name":"voteIdentifier",
            "in":"path",
            "required":true,
            "schema":{
              "type":"string"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/VotingDataResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components":{
    "schemas":{
      "VotingRequest":{
        "type":"object",
        "properties":{
          "voteIdentifier":{
            "type":"string"
          },
          "userId":{
            "type":"integer",
            "format":"int64"
          },
          "optionId":{
            "type":"integer",
            "format":"int64"
          }
        },
        "required":[
          "optionId",
          "userId",
          "voteIdentifier"
        ]
      },
...
      "VotingDataResponse":{
        "type":"object",
        "properties":{
          "voteIdentifier":{
            "type":"string"
          },
          "votingData":{
            "type":"object",
            "additionalProperties":{
              "type":"integer",
              "format":"int32"
            }
          }
        }
      },
...
    }
  }
}